backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
requests
numpy
pandas
scipy
python-jose[cryptography]
firebase-admin